requirements-test.txt
coverage>=4.0.2,<5
factory_boy
mock>=1.3.0,<4
tox>=2.2.1,<3
